feat: change food-record directory, footer
