+++ eatnfit-FE/src/components/AddForm/index.ts
@@ -1,0 +1,1 @@
+export { default as AddForm } from "./AddForm";
