+++ eatnfit-FE/src/components/AddedItems/index.ts
@@ -1,0 +1,1 @@
+export { default as AddedItems } from "./AddedItems";
