+++ eatnfit-FE/src/components/AddedItems/styles.ts
@@ -1,32 +1,4 @@
 import styled from "styled-components";
-import DeleteImg from "../../img/delete.png";
-
-type AddItemsProps = {
-  items: { name: string; calory: string }[];
-};
-
-const AddItems: React.FC<AddItemsProps> = (props) => {
-  return props.items.map((item) => (
-    <WrappedAddItems key={item.name}>
-      <Name>{item.name}</Name>
-      <Quantity>
-        <button type="button" className="minus">
-          -
-        </button>
-        <input type="text" className="inp" value="1" />
-        <button type="button" className="plus">
-          +
-        </button>
-      </Quantity>
-      <Calory>{item.calory}kcal</Calory>
-      <Delete>
-        <img src={DeleteImg} width="30px" />
-      </Delete>
-    </WrappedAddItems>
-  ));
-};
-
-export default AddItems;
 
 const WrappedAddItems = styled.div`
   border: 1px solid #ddd;
@@ -83,3 +55,5 @@ const Calory = styled.div`
 const Delete = styled.div`
   margin-left: auto;
 `;
+
+export { WrappedAddItems, Name, Quantity, Calory, Delete };
