+++ eatnfit-FE/src/components/CloseBtn/index.ts
@@ -1,0 +1,1 @@
+export { default as CloseBtn } from "./CloseBtn";
