+++ eatnfit-FE/src/components/Footer/Footer.tsx
@@ -1,5 +1,5 @@
+import { WrappedFooter } from "./styles";
 import footerPlus from "../../img/footerPlus.png";
-import styled from "styled-components";
 
 const Footer = () => {
   return (
@@ -10,9 +10,3 @@ const Footer = () => {
 };
 
 export default Footer;
-
-const WrappedFooter = styled.div`
-  padding: 8px;
-  display: flex;
-  justify-content: center;
-`;
