+++ eatnfit-FE/src/components/Footer/styles.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const WrappedFooter = styled.div`
+  padding: 8px;
+  display: flex;
+  justify-content: center;
+`;
+
+export { WrappedFooter };
