+++ eatnfit-FE/src/components/LongBtn/index.ts
@@ -1,0 +1,1 @@
+export { default as LongBtn } from "./LongBtn";
