+++ eatnfit-FE/src/components/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchInput } from "./SearchInput";
