+++ eatnfit-FE/src/components/SearchItems/SearchItems.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import { WrappedSearchItems, Calory, Image } from "./styles";
 import AddImg from "../../img/footerPlus.png";
 
 type SearchItemsProps = {
@@ -8,7 +8,7 @@ type SearchItemsProps = {
 const SearchItems: React.FC<SearchItemsProps> = (props) => {
   return props.items.map((item) => (
     <WrappedSearchItems key={item.name}>
-      <Name>{item.name}</Name>
+      <div className="name">{item.name}</div>
       <Calory>{item.calory}kcal, 1회 제공량</Calory>
       <Image src={AddImg} width="24px" />
       <hr />
@@ -17,21 +17,3 @@ const SearchItems: React.FC<SearchItemsProps> = (props) => {
 };
 
 export default SearchItems;
-
-const WrappedSearchItems = styled.div`
-  margin-top: 8px;
-  position: relative;
-`;
-
-const Name = styled.div``;
-const Calory = styled.div`
-  font-size: 14px;
-  color: gray;
-`;
-
-const Image = styled.img`
-  flex: 1;
-  position: absolute;
-  top: 20%;
-  right: 0;
-`;
