+++ eatnfit-FE/src/components/SearchItems/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchItems } from "./SearchItems";
