+++ eatnfit-FE/src/components/SearchItems/styles.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+const WrappedSearchItems = styled.div`
+  margin-top: 8px;
+  position: relative;
+`;
+
+const Calory = styled.div`
+  font-size: 14px;
+  color: gray;
+`;
+
+const Image = styled.img`
+  flex: 1;
+  position: absolute;
+  top: 20%;
+  right: 0;
+`;
+
+export { WrappedSearchItems, Calory, Image };
