+++ eatnfit-FE/src/components/SelectBtn/index.ts
@@ -1,0 +1,1 @@
+export { default as SelectBtn } from "./SelectBtn";
