+++ eatnfit-FE/src/components/SelectBtn/styles.ts
@@ -1,17 +1,5 @@
 import styled from "styled-components";
 
-type SelectBtnProps = {
-  items: string[];
-};
-
-const SelectBtn: React.FC<SelectBtnProps> = (props) => {
-  return props.items.map((item) => (
-    <WrappedSelectBtn key={item}>{item}</WrappedSelectBtn>
-  ));
-};
-
-export default SelectBtn;
-
 const WrappedSelectBtn = styled.button`
   width: 60px;
   height: 30px;
@@ -25,3 +13,5 @@ const WrappedSelectBtn = styled.button`
     background: #e0ebff;
   }
 `;
+
+export { WrappedSelectBtn };
