+++ eatnfit-FE/src/components/SubmitBtn/SubmitBtn.tsx
@@ -1,0 +1,7 @@
+import { WrappedSubmitBtn } from "./styltes";
+
+const SubmitBtn = () => {
+  return <WrappedSubmitBtn>등록</WrappedSubmitBtn>;
+};
+
+export default SubmitBtn;
