+++ eatnfit-FE/src/components/SubmitBtn/index.ts
@@ -1,0 +1,1 @@
+export { default as SubmitBtn } from "./SubmitBtn";
