+++ eatnfit-FE/src/components/SubmitBtn/styltes.ts
@@ -1,11 +1,5 @@
 import styled from "styled-components";
 
-const SubmitBtn = () => {
-  return <WrappedSubmitBtn>등록</WrappedSubmitBtn>;
-};
-
-export default SubmitBtn;
-
 const WrappedSubmitBtn = styled.button`
   width: 60px;
   height: 35px;
@@ -15,3 +9,5 @@ const WrappedSubmitBtn = styled.button`
   font-weight: 700;
   background: #e0ebff;
 `;
+
+export { WrappedSubmitBtn };
