+++ eatnfit-FE/src/components/index.ts
@@ -1,0 +1,9 @@
+export * from "./AddedItems";
+export * from "./AddForm";
+export * from "./CloseBtn";
+export * from "./Footer";
+export * from "./LongBtn";
+export * from "./SearchInput";
+export * from "./SearchItems";
+export * from "./SelectBtn";
+export * from "./SubmitBtn";
