+++ eatnfit-FE/src/pages/FoodRecordPage/FoodRecordPage.tsx
@@ -1,11 +1,24 @@
-import styled from "styled-components";
+import {
+  Wrap,
+  RecordHeader,
+  Main,
+  AddImg,
+  Category,
+  Time,
+  Calory,
+  Left,
+  Right,
+  ShowAddeditems,
+} from "./styles";
 import { Link } from "react-router-dom";
-import Close from "../../components/record/Close";
-import SubmitBtn from "../../components/record/SubmitBtn";
-import SelectBtn from "../../components/record/SelectBtn";
-import LongBtn from "../../components/record/LongBtn";
-import AddItems from "../../components/record/AddItems";
-import Footer from "../../components/common/Footer";
+import {
+  CloseBtn,
+  SubmitBtn,
+  SelectBtn,
+  LongBtn,
+  AddedItems,
+  Footer,
+} from "../../components";
 
 export default function FoodRecordPage() {
   const meal = ["아침", "아점", "점심", "간식", "점저", "저녁", "야식"];
@@ -17,7 +30,7 @@ export default function FoodRecordPage() {
   return (
     <Wrap>
       <RecordHeader>
-        <Close />
+        <CloseBtn />
         <SubmitBtn />
       </RecordHeader>
       <Main>
@@ -59,100 +72,12 @@ export default function FoodRecordPage() {
         <Link to="/foodrecord/search">
           <LongBtn text="+ 음식 검색하기" />
         </Link>
-        <ShowAdditems>
+        <ShowAddeditems>
           <span>추가한 음식</span>
-          <AddItems items={addItem} />
-        </ShowAdditems>
+          <AddedItems items={addItem} />
+        </ShowAddeditems>
       </Main>
       <Footer />
     </Wrap>
   );
 }
-
-const Wrap = styled.div`
-  max-height: 100vh;
-  overflow-y: auto;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-const RecordHeader = styled.div`
-  padding: 10px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const Main = styled.div`
-  padding: 0 15px;
-  h2 {
-    margin: 0 0 10px;
-  }
-  h4 {
-    margin: 10px 0;
-  }
-`;
-
-const AddImg = styled.div`
-  display: flex;
-  margin: 0 auto;
-  width: 200px;
-  height: 200px;
-  border: 1px solid gray;
-
-  & > span {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: gray;
-  }
-`;
-
-const Category = styled.div``;
-
-const Time = styled.div`
-  & > span {
-    color: gray;
-  }
-`;
-
-const Calory = styled.div`
-  h5 {
-    width: 90px;
-    margin: 0;
-  }
-  input {
-    width: 100px;
-    border-width: 0 0 1px;
-    text-align: center;
-    &:focus {
-      outline: none;
-    }
-    &:focus::placeholder {
-      color: transparent;
-    }
-  }
-  .first,
-  .second {
-    display: flex;
-    margin-bottom: 10px;
-  }
-`;
-
-const Left = styled.div`
-  width: 50%;
-  float: left;
-`;
-
-const Right = styled.div`
-  width: 50%;
-  float: right;
-`;
-
-const ShowAdditems = styled.div`
-  & > span {
-    font-size: 14px;
-  }
-`;
