+++ eatnfit-FE/src/pages/FoodRecordPage/SearchFoodPage.tsx
@@ -1,11 +1,19 @@
-import styled from "styled-components";
+import {
+  Wrap,
+  RecordHeader,
+  SearchFoodMain,
+  LinkToAddFood,
+  Items,
+} from "./styles";
 import { Link } from "react-router-dom";
-import Close from "../../components/record/Close";
-import SubmitBtn from "../../components/record/SubmitBtn";
-import SearchInput from "../../components/record/SearchInput";
-import LongBtn from "../../components/record/LongBtn";
-import SearchItems from "../../components/record/SearchItems";
-import Footer from "../../components/common/Footer";
+import {
+  CloseBtn,
+  SubmitBtn,
+  SearchInput,
+  LongBtn,
+  SearchItems,
+  Footer,
+} from "../../components";
 
 export default function SearchFoodPage() {
   const items = [
@@ -25,11 +33,11 @@ export default function SearchFoodPage() {
 
   return (
     <Wrap>
-      <SearchHeader>
-        <Close />
+      <RecordHeader>
+        <CloseBtn />
         <SubmitBtn />
-      </SearchHeader>
-      <Main>
+      </RecordHeader>
+      <SearchFoodMain>
         <h2>음식 검색</h2>
         <SearchInput text="음식" />
         <LinkToAddFood>
@@ -40,44 +48,8 @@ export default function SearchFoodPage() {
         <Items>
           <SearchItems items={items} />
         </Items>
-      </Main>
+      </SearchFoodMain>
       <Footer />
     </Wrap>
   );
 }
-
-const Wrap = styled.div`
-  max-height: 100vh;
-  overflow-y: auto;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-const SearchHeader = styled.div`
-  padding: 10px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const Main = styled.div`
-  & > h2 {
-    margin: 0 0 10px;
-    text-align: center;
-  }
-  padding: 0 28px;
-`;
-
-const LinkToAddFood = styled.div`
-  margin: 0 10px;
-`;
-
-const Items = styled.div`
-  height: 80vh;
-  margin: 4px 10px;
-  overflow-y: scroll;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
